requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
jinja2==3.1.4
groq
markupsafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pydub==0.25.1
pyyaml==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.14.0
setuptools==70.2.0
yt-dlp
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
typing-extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
diffusers==0.29.2
flask==3.0.3
flask-bootstrap==3.3.7.1
gdown==5.2.0
google-auth==2.31.0
google-api-python-client==2.136.0
accelerate==0.32.1
python-dotenv==1.0.0
